separare matches API from implementation

+++ include/TransformData.hpp
@@ -10,6 +10,6 @@ struct TransformInfo
     double  m_inlier_ratio = 0.0;
 };
 
-typedef std::pair<unsigned long, unsigned long> MatchIndex;
+typedef std::pair<unsigned int, unsigned int>   MatchIndex;
 typedef std::map<MatchIndex, TransformInfo>     TransformData;
 typedef std::pair<MatchIndex, TransformInfo>    TransformsEntry;
